Deactivate random deviation timeshift mode
The random deviation timeshift mode might lead to a situation in which
the jobs can be executed multiple times in the intended time range.

In order to fix this, we would need to track the vanilla planned time
and the make the executed run to sleep until the maximum deviation range
has passed. Example: planned schedule @midnight ±10min. The intended
launch time is somewhen from 23:50 until 00:10. Whenever the job is
launched in that range, it must be launched only once. If the job is
executed ~ 23:51, the used cron-library is calculating the next run.
Since @midnight has not passed yet, anywhere from 23:51 until 00:10
is a possible next launch time - and wrong.

This commit deactivates the "random timeshift deviation" mode for now.

The "random timeshift delay" mode is able to achieve _almost_ the same:
`50 23 * * * ~20m` is equal to the provided example of above with the
issue.